+++ packages/shared/src/components/cards/common/SquadPostCardHeader.tsx
@@ -4,7 +4,7 @@ import { Post } from '../../../graphql/posts';
 import { ProfileImageSize, ProfilePicture } from '../../ProfilePicture';
 import SourceButton from '../SourceButton';
 import PostMetadata from '../PostMetadata';
-import useBookmarkProvider from '../../../hooks/bookmark/useBookmarkProvider';
+import useBookmarkProvider from '../../../hooks/useBookmarkProvider';
 import {
   BookmakProviderHeader,
   headerHiddenClassName,
