+++ packages/shared/src/hooks/bookmark/index.ts
@@ -1,1 +1,2 @@
 export * from './useJustBookmarked';
+export * from './common';
